major updates to the laz.to.array function
updated the array.populator function to give the user a choice of using the pre-classified ground points or to take the lower point in each column of voxels so that a ground classification can be performed later.

Also, updated the noise removal part of the code so that it does not rely on classified points - instead in uses John Tukey's outlier identifier. This yields the same points being removed and allows users without classified point clouds to use this code.